Development

Former-commit-id: 32e40410ceaa33c50180fedcb98a1a935d9ddf85

+++ Intersect Server/Classes/Networking/PacketSender.cs
@@ -1144,8 +1144,8 @@ namespace Intersect.Server.Classes.Networking
         {
             var bf = new ByteBuffer();
             bf.WriteLong((int) ServerPackets.Experience);
-            bf.WriteInteger(client.Entity.Experience);
-            bf.WriteInteger(client.Entity.GetExperienceToNextLevel());
+            bf.WriteLong(client.Entity.Experience);
+            bf.WriteLong(client.Entity.GetExperienceToNextLevel());
             client.SendPacket(bf.ToArray());
             bf.Dispose();
         }
